fix healing not carrying to next level
GameSession was keeping a reference to the player's health since
we are re-creating the player each level.
The issue was that it was not following healing, so the health
would have gone back down at the start of the next level.
this could eventually cause negative health which would in turn cause errors.
To fix this I stopped following the player's health constantly and just copied
the health when completing the level which ensures it is the right value.
(could have also just follow healing, but what if we have effects and
such? lets not double store information if we can)
should possibly look into not needing this copy at all or making it
seemless